Initial commit (Laravel Blade): :tada: I start the section 3 of Laravel course
