+++ routes/web.php
@@ -1,7 +1,8 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\TestController;
+use Illuminate\Support\Facades\Route;
+
 
 /*
 |--------------------------------------------------------------------------
@@ -14,23 +15,4 @@ use App\Http\Controllers\TestController;
 |
 */
 
-Route::get('/',[TestController::class,'test']);
-
-Route::get('/custom', function () {
-
-    // SENDING PARAM TO custom.blade.php
-    $msg = "Message from server";
-    $anothermsg = "Anothe message from server";
-    // SENDING PARAM TO custom.blade.php VIA VARIABLE
-    $data = ['age' => 22];
-
-    return view('custom', ['msg' => $msg, 'anothermsg' => $anothermsg]);
-    // GIVING NAME TO THE ROUTE
-})->name('custom');
-
-
-Route::get('/contactme', function () {
-    return "Contact me";
-     // GIVING NAME TO THE ROUTE
-     // NAMES ALLOW THE ROUTES TO CHANGE
-})->name('contact');
+Route::get('/', [TestController::class, 'index']);
